Remove local-notifications plugin
The cordova-plugin-local-notifications-mm plugin is NOT working. Issuing local notification when app is in background or the screen is off does not occur.
However, local notifications do occur if a phone call is interrupted, but this does not really help. What is needed is issuing local notification when app is in the background. 
The config-file setting for parent = UIBackgroundModes did not help either so it is commented out too.