v1.1.022_20161215 for Distribution via HockeyApp
Commented out Screen Shot Report because it is working in HockeyApp
plugin (hockeyapp.composeFeedback(...)).
Leave Crash Test because it does work. However the dSym file is not
available for PhoneGap Build, so strack trace in a crash report is not
very user. Other info in the crash report is help helpful such as device
and IOS version.